improvement: implement Into trait for InputBoundingBox
